slow progress
